+++ src/components/common/index.ts
@@ -1,1 +1,2 @@
 export { default as GlobalNavigationBar } from './GlobalNavigationBar/GlobalNavigationBar';
+export { default as LinkMenuList } from './LinkMenuList/LinkMenuList';
